pytz: rework stubs
