--- stubs/pytz/@tests/stubtest_allowlist.txt
@@ -1,4 +1,0 @@
-pytz.BaseTzInfo.localize
-pytz.BaseTzInfo.normalize
-pytz.BaseTzInfo.zone
-pytz.FixedOffset
