+++ stubs/pytz/pytz/__init__.pyi
@@ -1,36 +1,35 @@
 import datetime
-from typing import Mapping
+from typing import ClassVar, Mapping
 
-class BaseTzInfo(datetime.tzinfo):
-    zone: str = ...
+from .exceptions import AmbiguousTimeError as AmbiguousTimeError
+from .exceptions import InvalidTimeError as InvalidTimeError
+from .exceptions import NonExistentTimeError as NonExistentTimeError
+from .exceptions import UnknownTimeZoneError as UnknownTimeZoneError
+from .tzinfo import BaseTzInfo as BaseTzInfo, DstTzInfo, StaticTzInfo
+
+# Actually named UTC and then masked with a singleton with the same name
+class _UTCclass(BaseTzInfo):
     def localize(self, dt: datetime.datetime, is_dst: bool | None = ...) -> datetime.datetime: ...
     def normalize(self, dt: datetime.datetime) -> datetime.datetime: ...
-
-class _UTCclass(BaseTzInfo):
     def tzname(self, dt: datetime.datetime | None) -> str: ...
     def utcoffset(self, dt: datetime.datetime | None) -> datetime.timedelta: ...
     def dst(self, dt: datetime.datetime | None) -> datetime.timedelta: ...
 
-class _StaticTzInfo(BaseTzInfo):
-    def tzname(self, dt: datetime.datetime | None, is_dst: bool | None = ...) -> str: ...
-    def utcoffset(self, dt: datetime.datetime | None, is_dst: bool | None = ...) -> datetime.timedelta: ...
-    def dst(self, dt: datetime.datetime | None, is_dst: bool | None = ...) -> datetime.timedelta: ...
-
-class _DstTzInfo(BaseTzInfo):
-    def tzname(self, dt: datetime.datetime | None, is_dst: bool | None = ...) -> str: ...
-    def utcoffset(self, dt: datetime.datetime | None, is_dst: bool | None = ...) -> datetime.timedelta | None: ...
-    def dst(self, dt: datetime.datetime | None, is_dst: bool | None = ...) -> datetime.timedelta | None: ...
-
-class UnknownTimeZoneError(KeyError): ...
-class InvalidTimeError(Exception): ...
-class AmbiguousTimeError(InvalidTimeError): ...
-class NonExistentTimeError(InvalidTimeError): ...
-
 utc: _UTCclass
 UTC: _UTCclass
 
-def timezone(zone: str) -> _UTCclass | _StaticTzInfo | _DstTzInfo: ...
-def FixedOffset(offset: int) -> _UTCclass | datetime.tzinfo: ...
+def timezone(zone: str) -> _UTCclass | StaticTzInfo | DstTzInfo: ...
+
+class _FixedOffset(datetime.tzinfo):
+    zone: ClassVar[None]
+    def __init__(self, minutes: int) -> None: ...
+    def utcoffset(self, dt: object) -> datetime.timedelta | None: ...
+    def dst(self, dt: object) -> datetime.timedelta: ...
+    def tzname(self, dt: object) -> None: ...
+    def localize(self, dt: datetime.datetime, is_dst: bool = ...) -> datetime.datetime: ...
+    def normalize(self, dt: datetime.datetime, is_dst: bool = ...) -> datetime.datetime: ...
+
+def FixedOffset(offset: int, _tzinfos: dict[int, _FixedOffset]) -> _UTCclass | _FixedOffset: ...
 
 all_timezones: list[str]
 all_timezones_set: set[str]
