+++ stubs/pytz/pytz/exceptions.pyi
@@ -1,0 +1,7 @@
+__all__ = ["UnknownTimeZoneError", "InvalidTimeError", "AmbiguousTimeError", "NonExistentTimeError"]
+
+class Error(Exception): ...
+class UnknownTimeZoneError(KeyError, Error): ...
+class InvalidTimeError(Error): ...
+class AmbiguousTimeError(InvalidTimeError): ...
+class NonExistentTimeError(InvalidTimeError): ...
