+++ stubs/pytz/pytz/tzinfo.pyi
@@ -1,0 +1,18 @@
+import datetime
+
+class BaseTzInfo(datetime.tzinfo):
+    zone: str  # Actually None but should be set on all concrete subclasses
+
+class StaticTzInfo(BaseTzInfo):
+    def localize(self, dt: datetime.datetime, is_dst: bool | None = ...) -> datetime.datetime: ...
+    def normalize(self, dt: datetime.datetime) -> datetime.datetime: ...
+    def tzname(self, dt: datetime.datetime | None, is_dst: bool | None = ...) -> str: ...
+    def utcoffset(self, dt: datetime.datetime | None, is_dst: bool | None = ...) -> datetime.timedelta: ...
+    def dst(self, dt: datetime.datetime | None, is_dst: bool | None = ...) -> datetime.timedelta: ...
+
+class DstTzInfo(BaseTzInfo):
+    def localize(self, dt: datetime.datetime, is_dst: bool | None = ...) -> datetime.datetime: ...
+    def normalize(self, dt: datetime.datetime) -> datetime.datetime: ...
+    def tzname(self, dt: datetime.datetime | None, is_dst: bool | None = ...) -> str: ...
+    def utcoffset(self, dt: datetime.datetime | None, is_dst: bool | None = ...) -> datetime.timedelta | None: ...
+    def dst(self, dt: datetime.datetime | None, is_dst: bool | None = ...) -> datetime.timedelta | None: ...
